+++ app/engine/shaders/terrain_entity_color_shader.vert
@@ -16,7 +16,8 @@ uniform float u_minZ;
 uniform float u_maxX;
 uniform float u_maxY;
 uniform float u_maxZ;
-uniform float u_textureRepeat;
+
+uniform int u_textureRepeat;
 
 uniform bool u_hasNormalMap;
 
@@ -36,7 +37,7 @@ void main()
 
 	f_worldSpacePos = worldSpacePosition.xyz;
 	f_shadowSpacePos = (u_shadowProjection * u_shadowView * worldSpacePosition);
-	f_uv = (v_uv * u_textureRepeat);
+	f_uv = (v_uv * float(u_textureRepeat));
 	f_normal = normalize(v_normal);
 	f_tbn = calculateTBN();
