+++ src/script_interpreter_func_fe3d_model_get.cpp
@@ -634,9 +634,9 @@ const bool ScriptInterpreter::_executeFe3dModelGetter(const string& functionName
 			{
 				if(_validateFe3dModelPart(args[0]->getString(), args[1]->getString()))
 				{
-					const auto result = _fe3d->model_getTextureRepeat(args[0]->getString(), args[1]->getString());
+					const auto result = static_cast<int>(_fe3d->model_getTextureRepeat(args[0]->getString(), args[1]->getString()));
 
-					returnValues.push_back(make_shared<ScriptValue>(SVT::DECIMAL, result));
+					returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 				}
 			}
 		}
