+++ src/script_interpreter_func_fe3d_terrain_get.cpp
@@ -414,9 +414,9 @@ const bool ScriptInterpreter::_executeFe3dTerrainGetter(const string& functionNa
 		{
 			if(_validateFe3dTerrain(args[0]->getString(), false))
 			{
-				const auto result = _fe3d->terrain_getTextureRepeat(_fe3d->terrain_getSelectedId());
+				const auto result = static_cast<int>(_fe3d->terrain_getTextureRepeat(_fe3d->terrain_getSelectedId()));
 
-				returnValues.push_back(make_shared<ScriptValue>(SVT::DECIMAL, result));
+				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
 		}
 	}
@@ -428,9 +428,9 @@ const bool ScriptInterpreter::_executeFe3dTerrainGetter(const string& functionNa
 		{
 			if(_validateFe3dTerrain(args[0]->getString(), false))
 			{
-				const auto result = _fe3d->terrain_getRedTextureRepeat(_fe3d->terrain_getSelectedId());
+				const auto result = static_cast<int>(_fe3d->terrain_getRedTextureRepeat(_fe3d->terrain_getSelectedId()));
 
-				returnValues.push_back(make_shared<ScriptValue>(SVT::DECIMAL, result));
+				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
 		}
 	}
@@ -442,9 +442,9 @@ const bool ScriptInterpreter::_executeFe3dTerrainGetter(const string& functionNa
 		{
 			if(_validateFe3dTerrain(args[0]->getString(), false))
 			{
-				const auto result = _fe3d->terrain_getGreenTextureRepeat(_fe3d->terrain_getSelectedId());
+				const auto result = static_cast<int>(_fe3d->terrain_getGreenTextureRepeat(_fe3d->terrain_getSelectedId()));
 
-				returnValues.push_back(make_shared<ScriptValue>(SVT::DECIMAL, result));
+				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
 		}
 	}
@@ -456,9 +456,9 @@ const bool ScriptInterpreter::_executeFe3dTerrainGetter(const string& functionNa
 		{
 			if(_validateFe3dTerrain(args[0]->getString(), false))
 			{
-				const auto result = _fe3d->terrain_getBlueTextureRepeat(_fe3d->terrain_getSelectedId());
+				const auto result = static_cast<int>(_fe3d->terrain_getBlueTextureRepeat(_fe3d->terrain_getSelectedId()));
 
-				returnValues.push_back(make_shared<ScriptValue>(SVT::DECIMAL, result));
+				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
 		}
 	}
