+++ src/script_interpreter_func_fe3d_water_get.cpp
@@ -287,9 +287,9 @@ const bool ScriptInterpreter::_executeFe3dWaterGetter(const string& functionName
 		{
 			if(_validateFe3dWater(args[0]->getString(), false))
 			{
-				const auto result = _fe3d->water_getTextureRepeat(_fe3d->water_getSelectedId());
+				const auto result = static_cast<int>(_fe3d->water_getTextureRepeat(_fe3d->water_getSelectedId()));
 
-				returnValues.push_back(make_shared<ScriptValue>(SVT::DECIMAL, result));
+				returnValues.push_back(make_shared<ScriptValue>(SVT::INTEGER, result));
 			}
 		}
 	}
